add to test suite and create likelihood module
