+++ tests/test_get_data.py
@@ -1,0 +1,5 @@
+""" Tests the data reading in routine """
+
+def test_get_obs():
+
+    return
