--- tests/test_mrprior.py
@@ -1,10 +1,0 @@
-""" test to check mr prior function operates correctly """
-from beans.mrprior import mr_prior
-
-def test_fit_file():
-    
-
-    M = 1.4
-    R = 11.2
-
-    return mr_prior(M, R)
