+++ tests/test_runmodel.py
@@ -1,10 +1,10 @@
 """ test to see if runmodel functions as expected """
 
-import beans.runmodel
+# import beans.runmodel
 
-def test_run_model():
+# def test_run_model():
 
-    test = runmodel(theta,y,tref,bstart,pflux, pfluxe, tobs,numburstssim,ref_ind,gti_checking)
+#     #test = runmodel(theta,y,tref,bstart,pflux, pfluxe, tobs,numburstssim,ref_ind,gti_checking)
 
-    assert 
-    return
+#     #assert 
+#     return
