+++ tests/test_settle.py
@@ -13,10 +13,7 @@ def test_settle_output():
 
     # run settle:
     res = settle.full(F=0.1,M=0.1,X=0.7,Z=0.02, C=0, R=11.2, Ma=1.4)
-    print(res)
-    res = settle.full(F=0.1,M=0.1,X=0.7,Z=0.02, C=0, R=11.2, Ma=1.4)
-    print(res)
-    result = np.allclose(res, [66.32432920153866, 7.1244386103642094, 7.516383459074593]) #alpha,tdel, E_b
+    result = np.allclose(res, [66.32432920153866, 4.630885096736736, 7.516383459074593]) #alpha,tdel, E_b
     assert result 
 
     return
